Move lin_binning and log_binning out of detail namespace into their own file
