+++ c++/triqs/stat/accumulator.hpp
@@ -19,6 +19,8 @@
 
 #pragma once
 
+#include "./lin_binning.hpp"
+#include "./log_binning.hpp"
 #include "./utils.hpp"
 #include "mean_error.hpp"
 #include <mpi/mpi.hpp>
@@ -35,231 +37,6 @@
 namespace triqs::stat {
 
   namespace details {
-
-    template <typename T> struct lin_binning {
-      long max_n_bins     = 0; // Maximum number of bins
-      long bin_capacity   = 1; // Current capacity of each bin (must be >= 1)
-      long last_bin_count = 0; // Number of data points the currently active bin [bins.back()]
-      std::vector<T> bins;     // Bins with accumulated data (stores means)
-      long count = 0;          // Total number of elements added to accumulator: Information only
-
-      // Callback function after compressing bins
-      using callback_t    = std::function<void(std::vector<T> const &)>;
-      callback_t callback = {};
-
-      // static std::string hdf5_format() { return "linear_bins"; }
-
-      lin_binning() = default;
-
-      lin_binning(T const &data_instance, long max_n_bins, long bin_capacity) : max_n_bins(max_n_bins), bin_capacity(bin_capacity) {
-        if (max_n_bins == 0) return;
-        T data_instance_local = data_instance;
-        data_instance_local   = 0;
-        if (max_n_bins > 0) { bins.reserve(max_n_bins); }
-        bins.emplace_back(std::move(data_instance_local));
-      }
-
-      // Optional constructor setting the callback function
-      lin_binning(T const &data_instance, long max_n_bins, long bin_capacity, callback_t const &callback)
-         : lin_binning(data_instance, max_n_bins, bin_capacity) {
-        this->callback = callback;
-      }
-
-      [[nodiscard]] long n_bins() const { return bins.size(); }
-
-      template <typename U> lin_binning<T> &operator<<(U &&x) {
-        ++count;
-        if (max_n_bins == 0) return *this;
-        // Check if all bins are full and compress if needed
-        if (max_n_bins > 1 && n_bins() == max_n_bins && last_bin_count == bin_capacity) {
-          if (callback) callback(bins); // Callback if defined
-          compress(2);                  // Adjusts bin_capacity & last_bin_count
-        }
-        // Check if current bin full: push new bin or add data to current bin
-        if (last_bin_count == bin_capacity && max_n_bins != 1) {
-          bins.emplace_back(std::forward<U>(x));
-          last_bin_count = 1;
-        } else {
-          last_bin_count++;
-          bins.back() += (x - bins.back()) / last_bin_count;
-        }
-        return *this;
-      }
-
-      // Compresses bins by scaling up bin_capacity by compression_factor (>= 2).
-      void compress(int compression_factor) {
-        if (max_n_bins == 0 || compression_factor < 2) return;
-        const int bins_left   = n_bins() % compression_factor;
-        int n_bins_new        = n_bins() / compression_factor;
-        int n_bins_last_chunk = compression_factor;
-        if (bins_left != 0) {
-          n_bins_new++;
-          n_bins_last_chunk = bins_left;
-        }
-        // Compress data into new bins, except the last new bin
-        for (int i = 0; i < n_bins_new - 1; ++i) {
-          if (i != 0) { bins[i] = std::move(bins[compression_factor * i]); }
-          for (int j = 1; j < compression_factor; j++) { bins[i] += bins[compression_factor * i + j]; }
-          bins[i] /= compression_factor;
-        }
-        // Last new bin is special: last old bin could be filled below capacity
-        int new_last_bin_count = last_bin_count + (n_bins_last_chunk - 1) * bin_capacity;
-        auto &new_last_bin     = bins[n_bins_new - 1];
-        new_last_bin           = std::move(bins[compression_factor * (n_bins_new - 1)]);
-        // If n_bins_last_chunk == 1, we have already copied its value above
-        if (n_bins_last_chunk > 1) {
-          for (int j = 1; j < n_bins_last_chunk - 1; j++) { new_last_bin += bins[compression_factor * (n_bins_new - 1) + j]; }
-          new_last_bin *= bin_capacity; // full bins in last chunk
-          new_last_bin += bins[compression_factor * (n_bins_new - 1) + (n_bins_last_chunk - 1)] * last_bin_count;
-          new_last_bin /= new_last_bin_count;
-        }
-        // Adjust final parameters
-        bins.resize(n_bins_new);
-        last_bin_count = new_last_bin_count;
-        bin_capacity *= compression_factor;
-      }
-    };
-
-    template <typename T> void h5_write(h5::group g, std::string const &name, lin_binning<T> const &l) {
-      auto gr = g.create_group(name);
-      h5_write(gr, "bins", l.bins);
-      h5_write(gr, "last_bin_count", l.last_bin_count);
-      h5_write(gr, "bin_capacity", l.bin_capacity);
-      h5_write(gr, "max_n_bins", l.max_n_bins);
-    }
-
-    template <typename T> void h5_read(h5::group g, std::string const &name, lin_binning<T> &l) {
-      auto gr = g.open_group(name);
-      h5_read(gr, "bins", l.bins);
-      h5_read(gr, "last_bin_count", l.last_bin_count);
-      h5_read(gr, "bin_capacity", l.bin_capacity);
-      h5_read(gr, "max_n_bins", l.max_n_bins);
-    }
-
-    // ************************************************************************************************
-    // Updating Formulae
-    // Mk \equiv \frac{1}{k} \sum_{i=1}^k x_i
-    // Qk \equiv \sum_{i=1}^k (x_i - M_k)^2
-    // Update formula
-    // M_{k} = M_{k-1} + \frac{x- M_{k-1}}{k}
-    // Q_{k} = Q_{k-1} + \frac{(k-1)(x- M_{k-1})^2}{k}
-    // Ref: see e.g. Chan, Golub, LeVeque. Am. Stat. 37, 242 (1983) and therein
-    template <typename T> struct log_binning {
-
-      std::vector<T> Mk; // Cf comments
-      using Q_t = get_real_t<T>;
-      std::vector<Q_t> Qk; // Cf comments
-      int max_n_bins = 0;
-      std::vector<T> acc;         // partial accumulators at size 2^(n+1). WARNING: acc[n] correspond to sum[n+1] to save unecessary temporaries
-      std::vector<int> acc_count; // number of elements for partial accumulators at size 2^(n+1)
-      long count = 0;             // Number of elements added to accumulator
-
-      //static std::string hdf5_format() { return "log_binning"; }
-
-      log_binning() = default;
-
-      log_binning(T const &data_instance, int max_n_bins) : max_n_bins(max_n_bins) {
-        if (max_n_bins == 0) return;
-        T data_instance_local = data_instance;
-        data_instance_local   = 0;
-        if (max_n_bins > 0) {
-          Qk.reserve(max_n_bins);
-          Mk.reserve(max_n_bins);
-          acc.reserve(max_n_bins - 1);
-          acc_count.reserve(max_n_bins - 1);
-        }
-        // If max_n_bins == 1, we don't need acc / acc_count, otherwise we initialize the first element
-        if (max_n_bins != 1) {
-          acc.emplace_back(data_instance_local);
-          acc_count.push_back(0);
-        }
-        // Multiply to ensure element can be element-wise multiplied [FIXME with concepts]
-        Qk.emplace_back(abs_square(data_instance_local));
-        Mk.emplace_back(std::move(data_instance_local));
-      }
-
-      [[nodiscard]] long n_bins() const { return Qk.size(); }
-
-      template <typename U> log_binning<T> &operator<<(U const &x) {
-        if (max_n_bins == 0) return *this;
-
-        using nda::conj;
-        using nda::real;
-
-        ++count;
-
-        // If max_n_bins == 1, there is only one (Mk, Qk) and we skip direclty to updating that below
-        if (max_n_bins != 1) {
-          // go up in n as long as the acc_count becomes full and add the acc in the
-          // then go down, and store the acc
-
-          int n = 0;
-          for (; n < acc.size(); ++n) {
-            if (n == 0) {
-              acc[n] += x; // n = 0 case is special, as it involves new data input
-            } else {
-              acc[n] += acc[n - 1];
-            }
-            acc_count[n]++;
-            if (acc_count[n] < 2) break;
-          }
-
-          // When reaching power of 2 nr data points, add a new pair of (Mk, Qk) = (acc.back(), 0)
-          // and add an extra space in acc, acc_count
-          if (count == (1 << acc.size()) && (max_n_bins < 0 || n_bins() < max_n_bins)) {
-            T last_acc = acc.back(); // Force copy
-
-            if (max_n_bins < 0 || n_bins() < (max_n_bins - 1)) {
-              acc.emplace_back(last_acc);
-              acc_count.push_back(1);
-            }
-
-            Mk.emplace_back(std::move(last_acc));
-            auto data_instance_q = Qk[0];
-            data_instance_q      = 0;
-            Qk.emplace_back(std::move(data_instance_q));
-          }
-
-          n--;
-          for (; n >= 0; n--) {
-            auto bin_capacity = (1ul << (n + 1));                    // 2^(n+1)
-            T x_m             = (acc[n] / bin_capacity - Mk[n + 1]); // Force T if expression template.
-            auto k            = count / bin_capacity;
-            Qk[n + 1] += static_cast<nda::get_value_t<Q_t>>((k - 1) / double(k)) * abs_square(x_m);
-            Mk[n + 1] += x_m / k;
-            acc_count[n] = 0;
-            acc[n]       = 0;
-          }
-        }
-
-        // Update the (Mk, Qk) pair with no binning (bin capacity: 2^0)
-        auto k = count;
-        T x_m  = (x - Mk[0]);
-        Qk[0] += static_cast<nda::get_value_t<Q_t>>((k - 1) / double(k)) * abs_square(x_m);
-        Mk[0] += x_m / k;
-
-        return *this;
-      }
-    };
-
-    // HDF5
-    template <typename T> void h5_write(h5::group g, std::string const &name, log_binning<T> const &l) {
-      auto gr = g.create_group(name);
-      h5_write(gr, "Qk", l.Qk);
-      h5_write(gr, "Mk", l.Mk);
-      h5_write(gr, "acc", l.acc);
-      h5_write(gr, "count", l.count);
-      h5_write(gr, "acc_count", l.acc_count);
-    }
-
-    template <typename T> void h5_read(h5::group g, std::string const &name, log_binning<T> &l) {
-      auto gr = g.open_group(name);
-      h5_read(gr, "Qk", l.Qk);
-      h5_read(gr, "Mk", l.Mk);
-      h5_read(gr, "acc", l.acc);
-      h5_read(gr, "count", l.count);
-      h5_read(gr, "acc_count", l.acc_count);
-    }
 
     template <typename T> auto mpi_reduce_MQ(const T &Mi, const get_real_t<T> &Qi, const long &count_i, mpi::communicator c, int root = 0) {
       using nda::conj;
@@ -319,8 +96,8 @@ namespace triqs::stat {
   template <typename T> class accumulator {
     private:
     long count = 0;
-    details::log_binning<T> log_bins;
-    details::lin_binning<T> lin_bins;
+    log_binning<T> log_bins;
+    lin_binning<T> lin_bins;
     std::vector<get_real_t<T>> auto_corr_times; // will be filled by the default callback if n_log_bins > 0
 
     // HDF5
