+++ test/c++/stat/mpi_MQ_reduce.cpp
@@ -26,7 +26,7 @@
 #include <cmath>
 #include <random>
 
-using namespace triqs::stat::details;
+using namespace triqs::stat;
 using nda::conj;
 
 mpi::communicator c;
@@ -46,7 +46,7 @@ TEST(Stat, ReduceSingleMQ) {
     if (c.size() * i / N == c.rank()) { logbin_i << x; }
   }
 
-  auto [M, Q, count_k] = mpi_reduce_MQ(logbin_i.Mk.at(0), logbin_i.Qk.at(0), logbin_i.count, c);
+  auto [M, Q, count_k] = details::mpi_reduce_MQ(logbin_i.Mk.at(0), logbin_i.Qk.at(0), logbin_i.count, c);
 
   if (c.rank() == 0) {
     EXPECT_NEAR(M, logbin_all.Mk.at(0), logbin_all.Mk.at(0) * std::numeric_limits<double>::epsilon() * 20);
@@ -54,7 +54,7 @@ TEST(Stat, ReduceSingleMQ) {
     EXPECT_EQ(count_k, logbin_all.count);
   }
 
-  auto [Mk1, Qk1, count_k1] = mpi_reduce_MQ(logbin_i.Mk.at(1), logbin_i.Qk.at(1), logbin_i.count >> 1, c);
+  auto [Mk1, Qk1, count_k1] = details::mpi_reduce_MQ(logbin_i.Mk.at(1), logbin_i.Qk.at(1), logbin_i.count >> 1, c);
 
   if (c.rank() == 0) {
     EXPECT_NEAR(Mk1, logbin_all.Mk.at(1), logbin_all.Mk.at(1) * std::numeric_limits<double>::epsilon() * 20);
